Add files via upload
this is my personal project named as jobjeevan , this project created for people who is searching for small type of jobs like labours , waiters & others. 
i use html , css & javascript language for creating this site. this project is not completed only designing & funcationality is added & backend development is still remaining. soon i will work on for backend devlopement on this project.